% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosomes.R
\name{chromosomes}
\alias{chromosomes}
\title{Mouse chromosomes}
\usage{
chromosomes(autosomal = TRUE, sexual = TRUE, mitochondrial = TRUE)
}
\arguments{
\item{autosomal}{Whether to include the autosomal chromosomes (1 thru 19).}

\item{sexual}{Whether to include the sexual chromosomes (X and Y).}

\item{mitochondrial}{Whether to include the mitochondrial chromosome (MT).}
}
\value{
A character vector of mouse chromosome names, or a subset thereof,
or an empty character vector.
}
\description{
\code{\link[=chromosomes]{chromosomes()}} returns mouse chromosome names.
}
\examples{
# All chromosomes.
chromosomes()

# Autosomal chromosomes.
chromosomes(autosomal = TRUE, sexual = FALSE, mitochondrial = FALSE)

}
