% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{common-args}
\alias{common-args}
\title{Common arguments}
\arguments{
\item{marker_id}{A character vector. MGI accession identifiers.}

\item{report_key}{A character vector. A key used to uniquely refer to an MGI
report.}

\item{report_file}{A character vector. The file path or URL to an MGI report
file.}

\item{report_type}{A character vector. The type of an MGI report.}
}
\description{
A set of arguments that are commonly reused across functions of
\code{{mgi.report.reader}}.
}
\keyword{internal}
