% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_types.R
\name{feature_types}
\alias{feature_types}
\title{Genome Feature types}
\usage{
feature_types()
}
\value{
A character vector of feature types' names.
}
\description{
\code{\link[=feature_types]{feature_types()}} returns different types of gene and genome features. For
feature type definitions, see \code{?feature_type_definitions}.
}
\examples{
feature_types()

}
