% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_mgi_identifier}
\alias{is_mgi_identifier}
\title{MGI accession identifier checking}
\usage{
is_mgi_identifier(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A logical vector.
}
\description{
\code{\link[=is_mgi_identifier]{is_mgi_identifier()}} checks whether the input is of the format
\code{"MGI:nnnnnn"}, where \code{"n"} is a digit and the number of digits can vary.
}
\keyword{internal}
