% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_types.R
\name{marker_types}
\alias{marker_types}
\title{Genetic marker types}
\usage{
marker_types()
}
\value{
A character vector.
}
\description{
\code{\link[=marker_types]{marker_types()}} returns MGI marker types.
}
\examples{
marker_types()

}
