% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgi.report.reader-package.R
\docType{package}
\name{mgi.report.reader-package}
\alias{mgi.report.reader}
\alias{mgi.report.reader-package}
\title{mgi.report.reader: MGI Report Readers}
\description{
Provides a set of functions designed to facilitate the reading, parsing, and analysis of MGI (Mouse Genome Informatics) report files: \url{https://www.informatics.jax.org/downloads/reports/index.html}.
}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Other contributors:
\itemize{
  \item Pattern Institute [copyright holder, funder]
}

}
\keyword{internal}
