% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.R
\name{mgi_base_url}
\alias{mgi_base_url}
\title{MGI base URL}
\usage{
mgi_base_url()
}
\value{
A scalar character vector with the URL.
}
\description{
\code{\link[=mgi_base_url]{mgi_base_url()}} returns MGI base URL.
}
\keyword{internal}
