% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.R
\name{mgi_reports_base_url}
\alias{mgi_reports_base_url}
\title{MGI reports base URL}
\usage{
mgi_reports_base_url()
}
\value{
A scalar character vector with the URL.
}
\description{
\code{\link[=mgi_reports_base_url]{mgi_reports_base_url()}} returns MGI reports base URL.
}
\keyword{internal}
