% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.R
\name{mgi_reports_index_url}
\alias{mgi_reports_index_url}
\title{MGI reports index URL}
\usage{
mgi_reports_index_url()
}
\value{
A scalar character vector with the URL.
}
\description{
\code{\link[=mgi_reports_index_url]{mgi_reports_index_url()}} returns MGI reports index URL.
}
\keyword{internal}
