% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_source.R
\name{report_datetime}
\alias{report_datetime}
\title{Report datetime}
\usage{
report_datetime(tbl)
}
\arguments{
\item{tbl}{Report data as a \link[tibble:tibble-package]{tibble}.}
}
\value{
A last modified date-time as a \link[base:DateTimeClasses]{POSIXct}
object.
}
\description{
\code{\link[=report_datetime]{report_datetime()}} the last modified date and time of the report source
(local file or remote file).
}
