% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{report_example}
\alias{report_example}
\title{Report example}
\usage{
report_example(report_file)
}
\arguments{
\item{report_file}{File basename.}
}
\description{
\code{\link[=report_example]{report_example()}} returns the local path of an example report file. These
files are typically very small and are useful for demonstrations. These
are mostly used in the Examples section of functions and in unit tests.
}
\examples{
report_example("MRK_List1-EX01.rpt")

report_example("MRK_List1-EX02.rpt")

report_example("MRK_List1-EX03.rpt")

}
