% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\docType{data}
\name{reports}
\alias{reports}
\title{Supported MGI reports}
\format{
A \link[tibble:tibble-package]{tibble} of 4 variables:

\describe{
\item{\code{report_key}}{A string key used to uniquely refer to an MGI report, which is only meaningful within the context of the \code{{mgi.report.reader}}.}
\item{\code{report_file}}{MGI report file name as hosted at \url{ https://www.informatics.jax.org/downloads/reports }.}
\item{\code{report_type}}{MGI report type. The type is used internally to find the appropriate reader for parsing, and is only meaningful within the context of \code{{mgi.report.reader}}.}
\item{\code{report_name}}{MGI report name. Report names are taken from \url{ https://www.informatics.jax.org/downloads/reports/index.html }.}
}
}
\usage{
reports
}
\description{
\code{reports} is a data set of supported MGI reports, meaning reports that
\code{{mgi.report.reader}} can currently read into R.
To browse all reports made available by MGI visit
\url{https://www.informatics.jax.org/downloads/reports}.
}
\examples{
reports

}
\keyword{datasets}
