% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feature_types}
\alias{feature_types}
\title{Genome Feature Type Definitions}
\format{
A \link[tibble:tibble-package]{tibble} with 71 rows
and 3 variables:
\describe{
\item{feature_type}{Character. The type of gene or genome feature.}
\item{so_id}{Character. The Sequence Ontology identifier associated with the feature type.}
\item{definition}{Character. The definition of the feature type.}
}
}
\source{
The table in \url{https://www.informatics.jax.org/userhelp/GENE_feature_types_help.shtml}
and a few other terms found in MGI reports.
}
\usage{
feature_types
}
\description{
A dataset containing different types of gene and genome features along with
their Sequence Ontology (SO) identifiers and definitions.
}
\examples{
print(feature_types, n = Inf)

}
\keyword{datasets}
