% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgi.report.reader-package.R
\docType{package}
\name{mgi.report.reader-package}
\alias{mgi.report.reader}
\alias{mgi.report.reader-package}
\title{mgi.report.reader: Read Mouse Genome Informatics Reports}
\description{
Provides readers for easy and consistent importing of Mouse Genome Informatics (MGI) report files: \url{https://www.informatics.jax.org/downloads/reports/index.html}. These data are provided by Baldarelli RM, Smith CL, Ringwald M, Richardson JE, Bult CJ, Mouse Genome Informatics Group (2024) \doi{10.1093/genetics/iyae031}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.pattern.institute/mgi.report.reader/}
  \item \url{https://github.com/patterninstitute/mgi.report.reader/}
  \item Report bugs at \url{https://github.com/patterninstitute/mgi.report.reader/issues}
}

}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Authors:
\itemize{
  \item David Shaw \email{David.Shaw@jax.org} (\href{https://orcid.org/0000-0002-9922-9934}{ORCID})
  \item Isabel Duarte \email{iduarte@pattern.institute} (\href{https://orcid.org/0000-0003-0060-2936}{ORCID})
  \item Ismail Gbadamosi \email{i.gbadamosi@nencki.edu.pl} (\href{https://orcid.org/0000-0002-1932-7519}{ORCID})
  \item Ali Jawaid \email{a.jawaid@nencki.edu.pl} (\href{https://orcid.org/0000-0002-5126-6744}{ORCID})
}

Other contributors:
\itemize{
  \item Nencki Institute of Experimental Biology [funder]
  \item University of Algarve [funder]
  \item The Jackson Laboratory [funder]
  \item Pattern Institute [copyright holder, funder]
}

}
\keyword{internal}
