% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgi.report.reader-package.R
\docType{package}
\name{mgi.report.reader-package}
\alias{mgi.report.reader}
\alias{mgi.report.reader-package}
\title{mgi.report.reader: MGI Report Readers}
\description{
Provides a set of functions designed to facilitate the reading, parsing, and analysis of MGI (Mouse Genome Informatics) report files: \url{https://www.informatics.jax.org/downloads/reports/index.html}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.pattern.institute/mgi.report.reader}
  \item \url{https://github.com/patterninstitute/mgi.report.reader}
  \item Report bugs at \url{https://www.pattern.institute/mgi.report.reader/issues}
}

}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Authors:
\itemize{
  \item Isabel Duarte \email{iduarte@pattern.institute} (\href{https://orcid.org/0000-0003-0060-2936}{ORCID})
  \item Ismail Gbadamosi \email{i.gbadamosi@nencki.edu.pl} (\href{https://orcid.org/0000-0002-1932-7519}{ORCID})
  \item Ali Jawaid \email{a.jawaid@nencki.edu.pl} (\href{https://orcid.org/0000-0002-1932-7519}{ORCID})
}

Other contributors:
\itemize{
  \item Nencki Institute of Experimental Biology [copyright holder, funder]
  \item University of Algarve [copyright holder, funder]
  \item Pattern Institute [copyright holder, funder]
}

}
\keyword{internal}
