% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_report.R
\name{read_report}
\alias{read_report}
\title{Read an MGI report}
\usage{
read_report(
  report_file,
  report_type = c("MRK_List1", "MRK_List2", "MGI_MRK_Coord", "MGI_Gene_Model_Coord",
    "MGI_GTGUP", "MRK_Sequence", "MRK_SwissProt_TrEMBL", "MRK_SwissProt", "MRK_GeneTrap",
    "MRK_ENSEMBL"),
  n_max = Inf
)
}
\arguments{
\item{report_file}{A path or URL to an MGI report file.}

\item{report_type}{Report type.}

\item{n_max}{Maximum number of lines to read.}
}
\value{
A \link[tibble:tibble-package]{tibble} with the report data in tidy
format.
}
\description{
\code{\link[=read_report]{read_report()}} reads in an MGI report.
}
\examples{
base_url <- "https://www.informatics.jax.org/downloads/reports"
n <- 10L

# Import the Mouse Genetic Markers (including withdrawn marker symbols) Report
read_report(file.path(base_url, "MRK_List1.rpt"), "MRK_List1", n_max = n)

# Import Mouse Genetic Markers (excluding withdrawn marker symbols) Report
# read_report(file.path(base_url, "MRK_List2.rpt"), "MRK_List2")

# Import the MGI Marker Coordinates' Report
# read_report(file.path(base_url, "MGI_MRK_Coord.rpt"), "MGI_MRK_Coord")

# Import the MGI Sequence Coordinates (in GFF format)
# read_report(file.path(base_url, "MGI_GTGUP.gff"), "MGI_GTGUP")

# Import the MGI Marker associations to Sequence (GenBank, RefSeq,Ensembl) information
# read_report(file.path(base_url, "MRK_Sequence.rpt"), "MRK_Sequence")

# Import MGI Marker associations to SWISS-PROT and TrEMBL protein IDs
# read_report(file.path(base_url, "MRK_SwissProt_TrEMBL.rpt"), "MRK_SwissProt_TrEMBL")

# Import MGI Marker associations to SWISS-PROT protein IDs
# read_report(file.path(base_url, "MRK_SwissProt.rpt"), "MRK_SwissProt")

# MGI Marker associations to Gene Trap IDs
# read_report(file.path(base_url, "MRK_GeneTrap.rpt"), "MRK_GeneTrap")

# MGI Marker associations to Ensembl sequence information
# read_report(file.path(base_url, "MRK_ENSEMBL.rpt"), "MRK_ENSEMBL")

}
