% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_report.R
\name{read_report}
\alias{read_report}
\title{Read an MGI report}
\usage{
read_report(
  report_file,
  report_type = c("MRK_List1", "MRK_List2", "MGI_MRK_Coord", "MGI_Gene_Model_Coord",
    "MGI_GTGUP")
)
}
\arguments{
\item{report_file}{A path or URL to an MGI report file.}

\item{report_type}{Report type.}
}
\value{
A \link[tibble:tibble-package]{tibble} with the report data in tidy
format.
}
\description{
\code{\link[=read_report]{read_report()}} reads in an MGI report.
}
\examples{
if (FALSE) {
base_url <- "https://www.informatics.jax.org/downloads/reports"

# Import the Mouse Genetic Markers (including withdrawn marker symbols) Report
read_report(file.path(base_url, "MRK_List1.rpt"), "MRK_List1")

# Import Mouse Genetic Markers (excluding withdrawn marker symbols) Report
read_report(file.path(base_url, "MRK_List2.rpt"), "MRK_List2")

# Import the MGI Marker Coordinates' Report
read_report(file.path(base_url, "MGI_MRK_Coord.rpt"), "MGI_MRK_Coord")

# Import the MGI Sequence Coordinates (in GFF format)
read_report(file.path(base_url, "MGI_GTGUP.gff"), "MGI_GTGUP")
}

}
